Echoing the changes made for the I/O version of the codelab.
