+++ ng2/1-skeleton/pubspec.yaml
@@ -4,7 +4,7 @@ version: 0.0.1
 environment:
   sdk: '>=1.13.0 <2.0.0'
 dependencies:
-  angular2: 2.0.0-beta.15
+  angular2: 2.0.0-beta.17
   browser: ^0.10.0
   dart_to_js_script_rewriter: ^1.0.1
 transformers:
