+++ ng2/2-blankbadge/lib/badge_component.html
