+++ ng2/3-inputnamebadge/lib/badge_component.css
