+++ ng2/3-inputnamebadge/lib/badge_component.dart
@@ -6,9 +6,9 @@ import 'package:angular2/core.dart';
 
 @Component(
     selector: 'pirate-badge',
-    templateUrl: 'pirate_badge_component.html',
-    styleUrls: const ['pirate_badge_component.css'])
-class PirateBadgeComponent {
+    templateUrl: 'badge_component.html',
+    styleUrls: const ['badge_component.css'])
+class BadgeComponent {
   String badgeName = '';
 
   void updateBadge(String inputName) {
