+++ ng2/4-buttonbadge/lib/badge_component.css
