+++ ng2/4-buttonbadge/lib/badge_component.dart
@@ -3,38 +3,28 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:angular2/core.dart';
-import 'pirate_name_service.dart';
 
 @Component(
     selector: 'pirate-badge',
-    templateUrl: 'pirate_badge_component.html',
-    styleUrls: const ['pirate_badge_component.css'],
-    providers: const [PirateNameService])
-class PirateBadgeComponent {
-  final PirateNameService _nameService;
+    templateUrl: 'badge_component.html',
+    styleUrls: const ['badge_component.css'])
+class BadgeComponent {
   String badgeName = '';
   String buttonText = 'Aye! Gimme a name!';
-  bool enableButton = true;
-
-  PirateBadgeComponent(this._nameService);
+  bool isButtonEnabled = true;
 
   void generateBadge() {
-    setBadgeName();
+    badgeName = 'Anne Bonney';
   }
 
   void updateBadge(String inputName) {
-    setBadgeName(inputName);
+    badgeName = inputName;
     if (inputName.trim().isEmpty) {
       buttonText = 'Aye! Gimme a name!';
-      enableButton = true;
+      isButtonEnabled = true;
     } else {
       buttonText = 'Arrr! Write yer name!';
-      enableButton = false;
+      isButtonEnabled = false;
     }
-  }
-
-  void setBadgeName([String newName = '']) {
-    if (newName == null) return;
-    badgeName = _nameService.getPirateName(newName);
   }
 }
