+++ ng2/5-piratenameservice/lib/badge_component.css
