+++ ng2/6-readjsonfile/lib/badge_component.css
